rename ptb-xl dataset
